+++ R/huePositionPlot.R
@@ -5,10 +5,19 @@
 #' @description A simple visualization of the hue positions for a given Munsell value/chroma according to Soil Survey Technical Note 2.
 #'
 #' @param value a single Munsell value
+#' 
 #' @param chroma a single Munsell chroma
+#' 
 #' @param chip.cex scaling for color chip rectangle
+#' 
 #' @param label.cex scaling for color chip 
-#' @param contour.dE00 logical, add dE00 contours from CIELAB coordinates (L,0,0), L is a constant value determined by `value` and `chroma`
+#' 
+#' @param contour.dE00 logical, add dE00 contours from `origin`, imlpicitly `TRUE` when `origin` is not `NULL` 
+#' 
+#' @param origin point used for distance comparisons can be either single row matrix of CIELAB coordinates, a character vector specifying a Munsell color. By default (`NULL`) represents CIELAB coordinates (L,0,0), where L is a constant value determined by `value` and `chroma`. See examples.
+#' 
+#' @param origin.cex scaling for origin point
+#' 
 #' @param grid.res grid resolution for contours, units are CIELAB A/B coordinates. Caution, small values result in many pair-wise distances which could take a very long time.
 #'
 #' @export
@@ -25,8 +34,22 @@
 #' 
 #' huePositionPlot(value = 6, chroma = 6, contour.dE00 = TRUE, grid.res = 2)
 #' 
+#' shift origin to arbitrary CIELAB coordinates or Munsell color
+#' huePositionPlot(origin = cbind(40, 5, 15), origin.cex = 0.5)
+#' 
+#' huePositionPlot(origin = '5G 6/4', origin.cex = 0.5)
+#' 
+#' huePositionPlot(origin = '10YR 3/4', origin.cex = 0.5)
+#' 
+#' huePositionPlot(value = 3, chroma = 4, origin = '10YR 3/4', origin.cex = 0.5)
 
-huePositionPlot <- function(value = 6, chroma = 6, chip.cex = 4.5, label.cex = 0.75, contour.dE00 = FALSE, grid.res = 2) {
+huePositionPlot <- function(value = 6, chroma = 6, chip.cex = 4.5, label.cex = 0.75, contour.dE00 = FALSE, origin = NULL, origin.cex = 0.75, grid.res = 2) {
+  
+  # interpret !NULL origin argument
+  if(!is.null(origin)) {
+    contour.dE00 <- TRUE
+  }
+    
   
   # dE00 contours requires farver >= 2.0.3
   if(contour.dE00) {
@@ -66,6 +89,8 @@ huePositionPlot <- function(value = 6, chroma = 6, chip.cex = 4.5, label.cex = 0
        axes = FALSE
   )
   
+  # add '0' lines
+  abline(h=0, v=0, lty=2)
   
   # add dE00 contours from [L,0,0]
   if(contour.dE00) {
@@ -86,9 +111,38 @@ huePositionPlot <- function(value = 6, chroma = 6, chip.cex = 4.5, label.cex = 0
       B = B.seq
     )
     
-    ## TODO: what would dE00 contours from some other point look like / mean?
-    # pair-wise distances from [L,0,0]
-    d <- farver::compare_colour(from = g, to = cbind(L.val, 0, 0), from_space = 'lab', to_space = 'lab', method = 'CIE2000')
+    # custom origin
+    if(!is.null(origin)) {
+      
+      # interpret origin
+      # matrix of CIELAB coordinates
+      if(inherits(origin, 'matrix')) {
+        # use as-is
+        .to_color <- origin
+        # stack coordinates for label
+        .to_label <- sprintf("%s\n%s\n%s", origin[1], origin[2], origin[3])
+        # sRGB color for display
+        .to_hex <- rgb(farver::convert_colour(.to_color, from = 'lab', to = 'rgb', white_to = 'D65') / 255)
+      }
+      
+      # munsell notation
+      if(inherits(origin, 'character')) {
+        # convert to CIELAB
+        .to_color <- as.matrix(parseMunsell(origin, returnLAB = TRUE), nrow = 1)
+        # split label into 2 lines
+        .pieces <- strsplit(origin, ' ', fixed = TRUE)[[1]]
+        .to_label <- sprintf("%s\n%s", .pieces[1], .pieces[2])
+        # sRGB color for display
+        .to_hex <- rgb(farver::convert_colour(.to_color, from = 'lab', to = 'rgb', white_to = 'D65') / 255)
+      }
+      
+      
+      # pair-wise distances from origin
+      d <- farver::compare_colour(from = g, to = .to_color, from_space = 'lab', to_space = 'lab', method = 'CIE2000')
+    } else {
+      # pair-wise distances from [L,0,0]
+      d <- farver::compare_colour(from = g, to = cbind(L.val, 0, 0), from_space = 'lab', to_space = 'lab', method = 'CIE2000')
+    }
     
     # init matrix of dE00 values along our grid of A and B coordinates
     m <- matrix(
@@ -106,15 +160,31 @@ huePositionPlot <- function(value = 6, chroma = 6, chip.cex = 4.5, label.cex = 0
       add = TRUE
     )
   }
-  
-  # add '0' lines
-  abline(h=0, v=0, lty=2)
-  
+
   # color chips
-  points(B ~ A, data=x, pch=15, col=x$hex, cex = chip.cex)
+  points(B ~ A, data = x, pch = 15, col = x$hex, cex = chip.cex)
   
   # label chips, color inversion based on legibility rules set by aqp::invertLabelColor
   text(x$A, x$B, labels = hue.labels, cex = label.cex, col = invertLabelColor(x$hex))
+  
+  # add optional origin point
+  if(!is.null(origin)) {
+    points(
+      x = .to_color[2], 
+      y = .to_color[3], 
+      pch = 15, 
+      col = .to_hex, 
+      cex = chip.cex
+    )
+    # annotate
+    text(
+      x = .to_color[2], 
+      y = .to_color[3], 
+      labels = .to_label, 
+      cex = origin.cex, 
+      col = invertLabelColor(.to_hex)
+    )
+  }
   
   # axis
   box()